added gradients_testutil files and fixed Status
