+++ tensorflow/c/eager/BUILD
@@ -229,6 +229,47 @@ tf_cuda_cc_test(
 )
 
 cc_library(
+    name = "gradients_testutil",
+    testonly = True,
+    srcs = [
+        "gradients_testutil.cc",
+    ],
+    hdrs = [
+         "gradients_testutil.h",
+    ],
+    visibility = [
+        "//tensorflow:internal",
+    ],
+    deps = [
+        ":abstract_context",
+        ":abstract_operation",
+        ":tape",
+        ":abstract_tensor_handle",
+        ":gradients",
+        ":c_api_experimental",
+        ":c_api_test_util",
+        ":c_api_unified_internal",
+        ":gradients_internal",
+        "//tensorflow/c/experimental/ops:array_ops",
+        "//tensorflow/c/experimental/ops:math_ops",
+        "//tensorflow/c/experimental/ops:nn_ops",
+        "//tensorflow/c:c_api",
+        "//tensorflow/c:c_test_util",
+        "//tensorflow/c:tf_status_helper",
+        "//tensorflow/cc/profiler",
+        "//tensorflow/compiler/mlir/tensorflow/c:mlir_c_api_registration",
+        "//tensorflow/core:lib",
+        "//tensorflow/core:protos_all_cc",
+        "//tensorflow/core:test",
+        "//tensorflow/core:test_main",
+        "//tensorflow/core/lib/llvm_rtti",
+        "@com_google_absl//absl/strings",
+        "@com_google_absl//absl/types:span",
+        "@com_google_absl//absl/container:flat_hash_map",
+    ],
+)
+
+cc_library(
     name = "mnist_gradients_testutil",
     srcs = [
         "mnist_gradients_testutil.cc",
@@ -273,6 +314,7 @@ cc_library(
         ":c_api_unified_internal",
         ":gradients_internal",
         ":mnist_gradients_testutil",
+        ":gradients_testutil",
         "//tensorflow/c:c_api",
         "//tensorflow/c:c_test_util",
         "//tensorflow/c:tf_status_helper",
