+++ tensorflow/c/eager/gradient_checker.h
@@ -12,8 +12,6 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-#include "tensorflow/c/eager/mnist_gradients_testutil.h"
-
 #include <memory>
 
 #include "absl/types/span.h"
@@ -24,6 +22,8 @@ limitations under the License.
 #include "tensorflow/c/eager/c_api_unified_experimental_internal.h"
 #include "tensorflow/c/eager/gradients.h"
 #include "tensorflow/c/eager/gradients_internal.h"
+#include "tensorflow/c/eager/gradients_testutil.h"
+#include "tensorflow/c/eager/mnist_gradients_testutil.h"
 #include "tensorflow/c/experimental/gradients/math_grad.h"
 #include "tensorflow/c/experimental/gradients/nn_grad.h"
 #include "tensorflow/c/experimental/ops/array_ops.h"
@@ -37,17 +37,19 @@ using Model = std::function<Status(
     AbstractContext*, absl::Span<AbstractTensorHandle* const>,
     absl::Span<AbstractTensorHandle*>, const GradientRegistry&)>;
 
-/** Returns numerical grad inside `dtheta_approx` given `forward` model and parameter
-  * specified by `gradIndex`
-  *
-  * `use_function` indicates whether to use graph mode(true) or eager(false)
-  *
-  * `is_scalar_out` should be true when `forward` returns a scalar TensorHandle; 
-  *  else GradientCheck will reduce_sum the tensor to get a scalar to estimate
-  *  the gradient with. Default is false. 
-  */
-Status GradientCheck(AbstractContext* ctx, Model forward, 
-                     std::vector<AbstractTensorHandle*> inputs,
-                     float* dtheta_approx, 
-                     int gradIndex, bool use_function,
-                     bool is_scalar_out=false);
+/** Returns numerical grad inside `dtheta_approx` given `forward` model and
+ * parameter specified by `input_index`.
+ *
+ * I.e. if y = <output of the forward model> and w = inputs[input_index],
+ * this will calculate dy/dw numerically.
+ *
+ * `use_function` indicates whether to use graph mode(true) or eager(false).
+ *
+ * `is_scalar_out` should be true when `forward` returns a scalar TensorHandle;
+ *  else this function will reduce_sum the tensor to get a scalar to estimate
+ *  the gradient with. Default is false.
+ */
+Status CalcNumericalGrad(AbstractContext* ctx, Model forward,
+                         std::vector<AbstractTensorHandle*> inputs,
+                         float* dtheta_approx, int input_index,
+                         bool use_function, bool is_scalar_out = false);
